+++ index.html
@@ -298,7 +298,7 @@
             integrity="sha384-Fy6S3B9q64WdZWQUiU+q4/2Lc9npb8tCaSX9FK7E8HnRr0Jz8D6OP9dO5Vg3Q9ct"
             crossorigin="anonymous"></script>
 
-        <script src="ble.js"></script>
+        <script type="module" src="js/ble.js"></script>
 
 </body>
