+++ js/ble.js
@@ -1,3 +1,6 @@
+import { gpio_asm_compile, set_gpio_asm_message, gpio_asm_file_read } from "./modules/gpioasm.js"
+import { set_device_battery_level, set_device_firmware_version, set_device_name, set_device_status } from "./modules/ui.js"
+
 const module = (function () {
     var is_device_connected = false
     var characteristic_configuration_pins = null
@@ -83,200 +86,6 @@ const module = (function () {
         display_pin_configuration_menu()
     }
 
-    function set_gpio_asm_message(message) {
-        $('#gpio-asm-file-info').text(message)
-    }
-
-    function gpio_asm_create_pin_encoder() {
-        return function (states_string) {
-            const states = []
-            for (var i = 0; i < states_string.length; i++) [
-                states.push({
-                    'h': true,
-                    '1': true,
-                    'l': false,
-                    '0': false,
-                    'i': undefined,
-                    'x': undefined,
-                    '-': undefined
-                }[states_string[i].toLowerCase()])
-            ]
-
-            return encode_states(states)
-        }
-    }
-
-    function gpio_asm_encode_uint_16(int_string) {
-        const value = Number(int_string)
-        const array = new Uint16Array(1)
-        array[0] = value
-        return new Uint8Array(array.buffer)
-    }
-
-    function gpio_asm_compile_command(command) {
-        const gpio_asm_encode_input_pins = gpio_asm_create_pin_encoder()
-        const gpio_asm_encode_output_pins = gpio_asm_create_pin_encoder()
-
-        function encode_pin_bytes_length(command, opcode_original) {
-            function get_bytes_count_needed_for_bits(state_string) {
-                return Math.ceil(state_string.length / 4)
-            }
-
-            const instruction = command.instruction
-            if (instruction == 'write_digital') {
-                return (opcode_original | get_bytes_count_needed_for_bits(command.args[0]))
-            }
-            if (instruction.startsWith('sleep_match')) {
-                return (opcode_original | get_bytes_count_needed_for_bits(command.args[0]))
-            }
-            if (instruction.startsWith('jump_match')) {
-                return (opcode_original | get_bytes_count_needed_for_bits(command.args[1]))
-            }
-        }
-
-        const gpio_asm_instrutions = {
-            write_digital: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b00000000,
-                argument_encoders: [
-                    gpio_asm_encode_output_pins
-                ]
-            },
-            write_analog_channel_0: {
-                instruction_bits: 0b00010000,
-                argument_encoders: [
-                    gpio_asm_encode_uint_16
-                ]
-            },
-            write_analog_channel_1: {
-                instruction_bits: 0b00010001,
-                argument_encoders: [
-                    gpio_asm_encode_uint_16
-                ]
-            },
-            write_analog_channel_2: {
-                instruction_bits: 0b00010010,
-                argument_encoders: [
-                    gpio_asm_encode_uint_16
-                ]
-            },
-            write_analog_channel_3: {
-                instruction_bits: 0b00010011,
-                argument_encoders: [
-                    gpio_asm_encode_uint_16
-                ]
-            },
-            sleep_ms: {
-                instruction_bits: 0b00100000,
-                argument_encoders: [
-                    encode_varint
-                ]
-            },
-            sleep_match_all: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b00110000,
-                argument_encoders: [
-                    gpio_asm_encode_input_pins
-                ]
-            },
-            sleep_match_any: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b01000000,
-                argument_encoders: [
-                    gpio_asm_encode_input_pins
-                ]
-            },
-            sleep_match_all_timeout: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b01010000,
-                argument_encoders: [
-                    gpio_asm_encode_input_pins,
-                    encode_varint
-                ]
-            },
-            sleep_match_any_timeout: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b01100000,
-                argument_encoders: [
-                    gpio_asm_encode_input_pins,
-                    encode_varint
-                ]
-            },
-            jump: {
-                instruction_bits: 0b01110000,
-                argument_encoders: [
-                    encode_varint
-                ]
-            },
-            check_bytecode_version: {
-                instruction_bits: 0b10000000,
-                argument_encoders: [
-                    encode_varint
-                ]
-            },
-            jump_match_all: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b10010000,
-                argument_encoders: [
-                    encode_varint,
-                    gpio_asm_encode_input_pins
-                ]
-            },
-            jump_match_any: {
-                opcode_encoder: encode_pin_bytes_length,
-                instruction_bits: 0b10100000,
-                argument_encoders: [
-                    encode_varint,
-                    gpio_asm_encode_input_pins
-                ]
-            },
-            jump_count: {
-                instruction_bits: 0b10110000,
-                argument_encoders: [
-                    encode_varint,
-                    encode_varint,
-                ]
-            },
-            exit: {
-                instruction_bits: 0b11000000,
-                argument_encoders: []
-            }
-        }
-
-        const instruction = command.instruction
-
-        if (instruction == 'label') {
-            // will take care of this later
-            return []
-        }
-
-        const instruction_info = gpio_asm_instrutions[instruction]
-
-        if (instruction_info == undefined) {
-            throw `instruction ${instruction} unknown`
-        }
-
-        const bytes = []
-
-        const opcode = function () {
-            if (instruction_info.opcode_encoder == undefined) {
-                return instruction_info.instruction_bits
-            }
-            return instruction_info.opcode_encoder(command, instruction_info.instruction_bits)
-        }()
-
-        bytes.push(opcode)
-
-        check_instruction_argument_length(command.instruction, instruction_info.argument_encoders.length, command.args.length)
-
-        for (var i = 0; i < command.args.length; i++) {
-            bytes.push(
-                ...instruction_info.argument_encoders[i](command.args[i])
-            )
-        }
-
-        return bytes
-    }
 
     function create_upload_packet_data_handler(characteristic) {
         return async function (data) {
@@ -297,7 +106,7 @@ const module = (function () {
     function on_button_gpio_asm_upload_click() {
         try {
             if (characteristic_gpio_asm_data == undefined) {
-                throw 'no gpioASM data characteristic found'
+                // throw 'no gpioASM data characteristic found'
             }
 
             gpio_asm_file_read(
@@ -310,217 +119,6 @@ const module = (function () {
         }
     }
 
-    function check_instruction_argument_length(instruction, expected_argument_count, actual_argument_count) {
-        if (actual_argument_count < expected_argument_count) {
-            throw `too few arguments for ${instruction}`
-        }
-        if (actual_argument_count > expected_argument_count) {
-            throw `too many arguments for ${instruction}`
-        }
-    }
-
-    function gpio_asm_compile(instructions) {
-        var current_offset = 0
-
-        var labels = {}
-
-        const bytecode_version = 0
-
-        instructions.splice(0, 0, {
-            instruction: 'check_bytecode_version',
-            args: [
-                bytecode_version
-            ]
-        })
-
-        for (const instruction of instructions) {
-            if (instruction.instruction == 'label') {
-                labels[instruction.args[0]] = current_offset
-                continue
-            }
-            instruction.offset = current_offset
-            if (instruction.instruction.startsWith('jump')) {
-                current_offset += 2
-                instruction.compiled_length = 2
-
-                if (instruction.instruction == 'jump_count') {
-                    check_instruction_argument_length('jump_count', 2, instruction.args.length)
-
-                    const length_count = encode_varint(Number(instruction.args[1])).length
-                    current_offset += length_count
-                    instruction.compiled_length += length_count
-                } else if (instruction.instruction.startsWith('jump_match_')) {
-                    check_instruction_argument_length(instruction.instruction, 2, instruction.args.length)
-
-                    const encoder = gpio_asm_create_pin_encoder()
-
-                    const length_count = encoder(instruction.args[1]).length
-                    current_offset += length_count
-                    instruction.compiled_length += length_count
-                } else {
-                    check_instruction_argument_length('jump', 1, instruction.args.length)
-                }
-
-                continue
-            }
-            var bytes = gpio_asm_compile_command(instruction)
-            current_offset += bytes.length
-            instruction.bytes = bytes
-        }
-
-        for (const instruction of instructions) {
-            if (!instruction.instruction.startsWith('jump')) {
-                continue
-            }
-            const target = instruction.args[0]
-            const address = labels[target]
-            if (address == undefined) {
-                throw `label ${target} unknown`
-            }
-
-            instruction.target_address = address
-
-            const args = [address]
-
-            if (instruction.instruction == 'jump_count') {
-                args.push(Number(instruction.args[1]))
-            } else if (instruction.instruction.startsWith('jump_match_')) {
-                args.push(instruction.args[1])
-            }
-
-            const bytes = gpio_asm_compile_command({
-                instruction: instruction.instruction,
-                args: args
-            })
-            instruction.bytes = bytes
-        }
-
-        while (true) {
-            // find first jump instruction with target address size wider than expected
-            const jump_mismatch_index = instructions.findIndex(instruction =>
-                instruction.instruction.startsWith('jump') &&
-                instruction.compiled_length != instruction.bytes.length
-            )
-            if (jump_mismatch_index == -1) {
-                // no mismatch found
-                break
-            }
-            if (jump_mismatch_index == instructions.length) {
-                // no instructions shifted since this is the last instruction
-                return
-            }
-            console.log(`resolving conflict with instruction ${jump_mismatch_index}`)
-
-            const jump_mismatch = instructions[jump_mismatch_index]
-            // calculate difference between expected and real address width
-            const length_dif = jump_mismatch.bytes.length - jump_mismatch.compiled_length
-
-            // calculate start off region that is getting shifted
-            const affected_address_start = instructions[jump_mismatch_index + 1].offset
-
-            // shift every following instructions offset by length dif
-            for (var i = (jump_mismatch_index + 1); i < instructions.length; i++) {
-                if (instructions[i].instruction == 'label') {
-                    continue
-                }
-                instructions[i].offset += length_dif
-            }
-
-            // adjust target addresses of all jump instruction
-            // that jump into the affected region
-            for (const instruction of instructions) {
-                if (!instruction.instruction.startsWith('jump')) {
-                    continue
-                }
-                var address = instruction.target_address
-                if (address < affected_address_start) {
-                    continue
-                }
-                address += length_dif
-                instruction.target_address = address
-
-                const args = [address]
-                if (instruction.instruction == 'jump_count') {
-                    args.push(Number(instruction.args[1]))
-                }
-
-                const bytes = gpio_asm_compile_command({
-                    instruction: instruction.instruction,
-                    args: args
-                })
-                instruction.bytes = bytes
-                instruction.compiled_length = bytes.length
-            }
-        }
-
-        const data = []
-
-        for (const instruction of instructions) {
-            if (instruction.instruction == 'label') {
-                continue
-            }
-
-            data.push(...instruction.bytes)
-        }
-
-        return data
-    }
-
-    function gpio_asm_file_read(compiled_data_handler) {
-        const files = $('#gpio-asm-file-upload')[0].files
-        if (files.length == 0) {
-            throw 'no file uploaded'
-        }
-        const file = files[0]
-        if (!file.name.endsWith('.gpioasm')) {
-            throw 'filename must end with .gpioasm'
-        }
-
-        const reader = new FileReader()
-        reader.onload = create_gpio_asm_file_load_handler(compiled_data_handler)
-        reader.readAsText(file)
-    }
-
-    function create_gpio_asm_file_load_handler(compiled_data_handler) {
-        return function (event) {
-            try {
-                const file_contents = event.currentTarget.result
-                const lines = file_contents.split('\n')
-                const commands = []
-                for (var line of lines) {
-                    line = line.trim()
-                    if (line == '') {
-                        continue
-                    }
-                    var command = line.match(/^[^ ]+ */)
-                    command = command[0]
-                    const argument_start = command.length
-                    var command = command.trim()
-
-                    var args = line.substring(argument_start)
-                    args = args.trim()
-
-                    if (args == '') {
-                        args = []
-                    } else {
-                        args = args.split(' ')
-                    }
-
-                    commands.push({
-                        instruction: command,
-                        args: args
-                    })
-                }
-
-                const data = gpio_asm_compile(commands)
-
-                console.log(data)
-                compiled_data_handler(data)
-            } catch (e) {
-                set_gpio_asm_message(e)
-            }
-        }
-    }
 
     function handle_gpio_asm_upload(event) {
 
@@ -847,143 +445,6 @@ const module = (function () {
                 'b1190000-2a74-d5a2-784f-c1cdb3862ab0' // gpioASM service
             ]
         }).then(on_bluetooth_device_selected)
-    }
-
-    function filter_states(states) {
-        if (states.length == 0) {
-            return []
-        }
-        if (states.length == 1) {
-            return states
-        }
-        const last_state = [...states[0]]
-        const filtered_states = [
-            [...states[0]]
-        ]
-
-        for (var i = 1; i < states.length; i++) {
-            const state = states[i]
-            const filtered_state = []
-
-            for (var j = 0; j < state.length; j++) {
-                const is_high = state[j]
-                if (is_high == last_state[j]) {
-                    filtered_state.push(undefined)
-                    continue
-                }
-                filtered_state.push(is_high)
-                last_state[j] = is_high
-            }
-            filtered_states.push(filtered_state)
-        }
-
-        return filtered_states
-    }
-
-    function encode_states(states) {
-        const bytes_needed = Math.ceil(states.length / 4)
-        const bytes = Array(bytes_needed).fill(0xff)
-        for (var i = 0; i < states.length; i++) {
-            const state = states[i]
-            if (state == undefined) {
-                continue
-            }
-
-            const byte_index = Math.floor(i / 4)
-            const bit_shift = 6 - ((i * 2) % 8)
-
-            if (state) {
-                bytes[byte_index] &= ~(0b10 << bit_shift)
-            } else {
-                bytes[byte_index] &= ~(0b11 << bit_shift)
-            }
-        }
-
-        return bytes
-    }
-
-    function encode_analog_values(analog_values) {
-        analog_values = analog_values.map(value => (value == undefined) ? 0xffff : value)
-        const array = new Uint16Array(analog_values)
-        return new Uint8Array(array.buffer)
-    }
-
-    function encode_varint(number) {
-        var bits_needed = 1
-        if (number != 0) {
-            bits_needed = Math.floor(Math.log2(number)) + 1
-        }
-        const bit_mask = 1 << (bits_needed - 1)
-        const bits = []
-        for (var i = 0; i < bits_needed; i++) {
-            bits.push((number & (bit_mask >> i)) != 0)
-        }
-        const bytes = []
-        const bytes_needed = Math.ceil(bits.length / 7)
-
-        for (var i = 0; i < bytes_needed; i++) {
-            var byte = 0x00
-            const start_bit_index = bits.length - (i * 7)
-            for (var j = 0; j < 7; j++) {
-                byte |= bits[start_bit_index - j - 1] << j
-            }
-            bytes.push(byte)
-        }
-
-        for (var i = 0; i < (bytes.length - 1); i++) {
-            bytes[i] |= 0b10000000
-        }
-
-        return bytes
-    }
-
-    function encode_pin(pin) {
-        var byte = 0b1111
-
-        if (pin == undefined) {
-            return byte;
-        }
-
-        if (pin.function == 'output') {
-            byte = 0b0000
-            if (pin.function_output == 'analog') {
-                byte |= 0b0100
-            } else if (pin.default_high) {
-                byte |= 0b0010
-            }
-        } else if (pin.function == 'input') {
-            byte = 0b1000
-            if (pin.pull == 'pullup') {
-                byte |= 0b0010
-            } else if (pin.pull == 'pulldown') {
-                byte |= 0b0100
-            }
-        }
-
-        if (pin.invert) {
-            byte |= 0b0001
-        }
-        return byte;
-    }
-
-    function encode_two_pins(pin0, pin1) {
-        return (encode_pin(pin0) << 0) | (encode_pin(pin1) << 4);
-    }
-
-    function encode_pin_configuration(configured_pins) {
-        const bytes = []
-        for (var i = 0; i < Math.ceil(configured_pins.length / 2); i++) {
-            bytes.splice(0, 0, encode_two_pins(
-                configured_pins[(i * 2) + 0],
-                configured_pins[(i * 2) + 1]
-            ))
-        }
-        // disable pin is count is uneven
-        if ((configured_pins.length % 2) == 1) {
-            bytes[0] |= 0b11110000
-        }
-
-        return bytes
     }
 
     async function on_button_send_configuraion_click(_) {
@@ -2059,22 +1520,6 @@ const module = (function () {
         if (advertising_interval != 0xffff) {
             $('#input-advertising-interval').val(advertising_interval)
         }
-    }
-
-    function set_device_status(status) {
-        $('#info-device-status').text(`Device status: ${status}`)
-    }
-
-    function set_device_name(name) {
-        $('#info-device-name').text(`Device name: ${name}`)
-    }
-
-    function set_device_battery_level(level) {
-        $('#info-device-battery-level').text(`Device battery level: ${level}%`)
-    }
-
-    function set_device_firmware_version(version) {
-        $('#info-device-firmware-version').text(`Device firmware version: ${version}`)
     }
 
     return {
